+++ src/id_provider/id_provider_mutex.rs
@@ -6,17 +6,17 @@ use super::IdProvider;
 /// This is probably too slow to be of any practical use; for test only
 #[derive(Debug, Default)]
 pub struct IdProviderMutex {
-    internal_counter: Mutex<u64>,
+    internal_counter: Mutex<usize>,
 }
 
 impl PartialEq for IdProviderMutex {
-    fn eq(&self, _other: &Self) -> bool {
-        todo!()
+    fn eq(&self, other: &Self) -> bool {
+        self.peek_next_id() == other.peek_next_id()
     }
 }
 
 impl IdProviderMutex {
-    pub fn new(internal_counter: u64) -> Self {
+    pub fn new(internal_counter: usize) -> Self {
         Self {
             internal_counter: Mutex::new(internal_counter),
         }
@@ -30,18 +30,29 @@ impl Clone for IdProviderMutex {
 }
 
 impl IdProvider for IdProviderMutex {
-    fn get_next_id(&self) -> u64 {
+    fn get_next_id(&self) -> usize {
         let mut counter = self.internal_counter.lock().unwrap();
         // copy the value before mutating
         let value = *counter;
         *counter += 1;
         value
     }
+    fn peek_next_id(&self) -> usize {
+        *self.internal_counter.lock().unwrap()
+    }
 }
 
 #[cfg(test)]
 mod tests {
     use crate::id_provider::{id_provider_mutex::IdProviderMutex, IdProvider};
+
+    #[test]
+    fn peek_id_as_expected() {
+        let target = IdProviderMutex::new(5);
+
+        assert_eq!(5, target.peek_next_id());
+        assert_eq!(5, target.peek_next_id());
+    }
 
     #[test]
     fn gets_ids_as_expected() {