+++ src/id_provider/mod.rs
@@ -10,5 +10,6 @@ pub mod sized_id_provider;
 /// It needs to implement send and sync so that it can be safely passed between threads. \
 /// It is included for use in the examples
 pub trait IdProvider: Debug + Send + Sync {
-    fn get_next_id(&self) -> u64;
+    fn peek_next_id(&self) -> usize;
+    fn get_next_id(&self) -> usize;
 }